third_party/WebKit/LayoutTests/fast/dom/Window/lenient-this-issue569043-expected.txt


Test that Issue 569043 has been fixed and [LenientThis] and 'super' are working as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.onmousedown is Function
PASS window.onmouseenter is Function
PASS frame1.onmousedown is null
PASS frame1.onmouseenter is null
PASS successfullyParsed is true

TEST COMPLETE
